requirements.txt
RPi.GPIO==0.6.3
argparse==1.2.1
beautifulsoup4==4.5.3
requests==2.13.0
wsgiref==0.1.2
